undo repr
