+++ qiskit/circuit/quantumcircuit.py
@@ -36,8 +36,6 @@ from .bit import Bit
 def _is_bit(obj):
     """Determine if obj is a bit"""
     # If there is a bit type this could be replaced by isinstance.
-    if isinstance(obj, Bit):
-        return False
     if isinstance(obj, tuple) and len(obj) == 2:
         if isinstance(obj[0], Register) and isinstance(obj[1], int) and obj[1] < len(obj[0]):
             warn('Referring to a bit as a tuple is being deprecated. '
@@ -273,12 +271,12 @@ class QuantumCircuit:
     def _bit_argument_conversion(bit_representation, in_array):
         ret = None
         try:
-            if _is_bit(bit_representation):
-                # circuit.h((qr, 0)) -> circuit.h([qr[0]])
-                ret = [bit_representation[0][bit_representation[1]]]
-            elif isinstance(bit_representation, Bit):
+            if isinstance(bit_representation, Bit):
                 # circuit.h(qr[0]) -> circuit.h([qr[0]])
                 ret = [bit_representation]
+            elif _is_bit(bit_representation):
+                # circuit.h((qr, 0)) -> circuit.h([qr[0]])
+                ret = [bit_representation[0][bit_representation[1]]]
             elif isinstance(bit_representation, Register):
                 # circuit.h(qr) -> circuit.h([qr[0], qr[1]])
                 ret = bit_representation[:]
