+++ qiskit/circuit/quantumregister.py
@@ -16,6 +16,8 @@
 Quantum register reference object.
 """
 import itertools
+from warnings import warn
+import inspect
 
 from qiskit.exceptions import QiskitError
 from .register import Register
@@ -32,9 +34,16 @@ class Qubit(Bit):
             raise QiskitError('Qubit needs a QuantumRegister and %s was provided' %
                               type(register).__name__)
 
-    def __repr__(self):
-        """Return the official string representing the qubit."""
-        return "Qubit(%s, %s)" % (self.register, self.index)
+    @property
+    def __class__(self):
+        curframe = inspect.currentframe()
+        calframe = inspect.getouterframes(curframe, 2)
+        for context in calframe[1].code_context:
+            if 'tuple' in context:
+                warn('Bit-as-tuple is deprecated. Replace isinstance(qr[0], tuple) for '
+                     'isinstance(qr[0], Qubit).', DeprecationWarning, stacklevel=2)
+                return tuple
+        return Qubit
 
 
 class QuantumRegister(Register):
